corecciones rami

+++ src/router/carts.js
@@ -13,32 +13,44 @@ router.post("", async (req, res) => {
 
 router.delete("/:cid", async (req, res) => {
   const cart = new CartDAO();
-  const borrar = await cart.deleteById(req.params.cid);
-  res.send(`Se borro ${borrar}`);
+  const carrito = await cart.getAll(req.params.cid);
+  let cartExiste = await cart.getById(carrito);
+  if (cartExiste != undefined) {
+    const borrar = await cart.deleteById(req.params.cid);
+    res.send(`Se borro ${borrar}`);
+  } else {
+    res.send("Carrito erroneo");
+  }
 });
 
 router.get("/:cid/products", async (req, res) => {
   const cart = new CartDAO();
   const carrito = await cart.getAll(req.params.cid);
-  res.send({ cart: carrito });
+  let cartExiste = await cart.getById(carrito);
+  console.log(cartExiste);
+  if (cartExiste != undefined) {
+    res.send({ cart: carrito });
+  } else {
+    res.send("Carrito erroneo");
+  }
 });
 
 router.post("/:cid/products", async (req, res) => {
   let idsProduct = req.body.id;
   const prod = new ProductDAO();
-  for (let idProducto of idsProduct) {
-    const items = await prod.getAll(idProducto);
-    const itemElegido = items[0];
-    if (itemElegido.stock > 0) {
-      itemElegido.stock--;
-      await prod.changeById(itemElegido.id, itemElegido);
-
-      const cart = new CartDAO();
-      let carrito = await cart.getAll(req.params.cid);
-      let carritoNew = carrito[0].producto.find((p) => p.id == idProducto);
+  const items = await prod.getAll(idsProduct);
+  const itemElegido = await prod.getById(items[0]._id);
+  if (itemElegido.stock > 0) {
+    itemElegido.stock--;
+    await prod.changeById(itemElegido.id, itemElegido);
+    const cart = new CartDAO();
+    let carrito = await cart.getAll(req.params.cid);
+    let cartExiste = await cart.getById(carrito);
+    if (cartExiste != undefined) {
+      let carritoNew = carrito[0].producto.find((p) => p.id == idsProduct);
       if (carritoNew != undefined) {
         carritoNew.quantity = carritoNew.quantity + 1;
-        let carritoMod = carrito[0].producto.filter((p) => p.id != idProducto);
+        let carritoMod = carrito[0].producto.filter((p) => p.id != idsProduct);
         carritoMod.push(carritoNew);
         carritoNew = carritoMod;
 
@@ -47,11 +59,13 @@ router.post("/:cid/products", async (req, res) => {
         delete itemElegido._doc.stock;
         await cart.saveProduct(itemElegido, req.params.cid);
       }
+      res.send("se agrego");
     } else {
-      res.send(`No hay stock de ${itemElegido.id}`);
+      res.send("Carrito erroneo");
     }
+  } else {
+    res.send(`No hay stock de ${itemElegido.id}`);
   }
-  res.send("se agrego");
 });
 
 router.delete("/:cid/products/:pid", async (req, res) => {
@@ -61,25 +75,29 @@ router.delete("/:cid/products/:pid", async (req, res) => {
   let itemElegido = items[0];
   const cart = new CartDAO();
   let carrito = await cart.getAll(req.params.cid);
-  let carritoNew = carrito[0].producto.find((p) => p.id == req.params.pid);
-  if (carritoNew != undefined) {
-    itemElegido.stock++;
-    await prod.changeById(itemElegido.id, itemElegido);
+  let cartExiste = await cart.getById(carrito);
+  if (cartExiste != undefined) {
+    let carritoNew = carrito[0].producto.find((p) => p.id == req.params.pid);
+    if (carritoNew != undefined) {
+      itemElegido.stock++;
+      await prod.changeById(itemElegido.id, itemElegido);
 
-    if (carritoNew.quantity > 1) {
-      carritoNew.quantity = carritoNew.quantity - 1;
-      let carritoMod = carrito[0].producto.filter(
-        (p) => p.id != req.params.pid
-      );
-      carritoMod.push(carritoNew);
-      carritoNew = carritoMod;
-      console.log(carritoNew);
-    } else {
-      carritoNew = carrito[0].producto.filter((p) => p.id != req.params.pid);
+      if (carritoNew.quantity > 1) {
+        carritoNew.quantity = carritoNew.quantity - 1;
+        let carritoMod = carrito[0].producto.filter(
+          (p) => p.id != req.params.pid
+        );
+        carritoMod.push(carritoNew);
+        carritoNew = carritoMod;
+      } else {
+        carritoNew = carrito[0].producto.filter((p) => p.id != req.params.pid);
+      }
     }
+    await cart.deleteUpdateProduct(req.params.cid, carritoNew);
+    res.send("se borro");
+  } else {
+    res.send("Carrito erroneo");
   }
-  await cart.deleteUpdateProduct(req.params.cid, carritoNew);
-  res.send("se borro");
 });
 
 export default router;
